x/sigmoidtest/types/message_approve_request_test.go
package types

import (
	"testing"

	"sigmoid-test/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgApproveRequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgApproveRequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgApproveRequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgApproveRequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
